fixed svg visibility and play buttons
now the listeners need to implement a play and a pause functions.
these are called when the state is played or pause. When the state
is paused then the tick function is no longer called. Therefore
listeners like that of a video player need to implement play and pause
to take appropriate actions