Final (?) debugging of angle calculations + change of input format
Now the calculations of angles should be correct. Made the code more
efficient by placing some redundant code in new functions. Also changed
the input so that each joint is just a simple array instead of an object
with x, y and z properties. It just made it more difficult when using
the vector calculations.